backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
psycopg[binary]>=3.1.18
celery[redis]>=5.3.6
django-celery-beat>=2.6.0
gunicorn>=22.0.0
pytest>=8.2.0
pytest-django>=4.8.0
django-cors-headers>=4.4.0
